To -Do List
Discription

Problem statement
Create a simple Todo list with vanilla JS (as you see in the GIF below). Vanilla JS means no plugins, no libraries, nothing.

Instructions
You are ALLOWED to style the todo list the way you like. Do whatever you want with the styling.
You are ALLOWED to google and read how to approach the problem.
DO NOT copy and paste code from the internet.
DO NOT cheat with other students. REMEMBER: These tests are for YOU so that you can learn something new. Cheating won’t help. It’s better to not submit the solution than to cheat and submit. PS: There will be a plagiarism check and when found cheating, you won’t be allowed to give future tests.
Once you have finished with the test, make a video recording your computer screen (via phone or a software like OBS/screencastify etc) and explaining how you have approached the problem in code and showing the final product. The video can be UPTO 5 mins long. Please don’t record videos longer than that.
Upload and Host the code on github as well.
Once you have recorded the video, upload it on youtube (unlisted or whatever), drive or wherever you want to upload it and submit the link.
Hosting on Github is compulsory, I will be rejecting the submission if not hosted!

Features (No need for extra features, just make the listed features)
A minimal design with HTML and CSS, it should look nice (no libraries, no frameworks).
A text box at the top to add a todo item to the list.
Add button: when a user clicks on the button, add that todo item to the list.
A delete button to delete the todo item completely from the list.
A checkbox to check the item off the list. (checked item should stay in the list, with some differentiation from rest of the todo items)
Total number of tasks in the list.
Extra points for creativity (think about user accessibility, some animations and things like that)